+++ footer.php
@@ -30,6 +30,7 @@
         </div>
     </div>
 </div>
+
 <div class="modal hide fade do-not-print modal-gallery" id="modal-gallery" tabindex="-1" aria-hidden="true">
   <div class="modal-header">
     <button type="button" class="close" data-dismiss="modal" aria-hidden="true">&times;</button>
@@ -39,10 +40,10 @@
     <div class="modal-image"></div>
   </div>
   <div class="modal-footer">
-    <a class="btn btn-primary modal-next">Next <i class="icon-arrow-right icon-white"></i></a>
-    <a class="btn btn-info modal-prev"><i class="icon-arrow-left icon-white"></i> Previous</a>
-    <a class="btn btn-success modal-play modal-slideshow" data-slideshow="5000"><i class="icon-play icon-white"></i> Slideshow</a>
-    <a class="btn modal-download" target="_blank"><i class="icon-download"></i> Download</a>
+    <a class="btn btn-primary modal-next"><?php _e("Suivant","stormbringer"); ?> <i class="icon-arrow-right icon-white"></i></a>
+    <a class="btn btn-info modal-prev"><i class="icon-arrow-left icon-white"></i> <?php _e("Précédent","stormbringer"); ?></a>
+    <a class="btn btn-success modal-play modal-slideshow" data-slideshow="5000"><i class="icon-play icon-white"></i> <?php _e("Diaporama","stormbringer"); ?></a>
+    <!--<a class="btn modal-download" target="_blank"><i class="icon-download"></i> Download</a>-->
   </div>
 </div>
 <!-- /#modal-gallery -->