+++ js/app.js
@@ -1,23 +1,3 @@
-// Fancybox
-if(jQuery().fancybox)
-{
-  // Fancybox : default config
-  $.extend($.fn.fancybox.defaults, {
-    padding : 20,
-    margin :20,
-    width : '50%',
-    height : '50%',
-    overlayColor	:	'#333333',
-    titlePosition	:	'inside',
-    cyclic		:	true,
-    autoScale		:	false,
-    centerOnScroll		:	true,
-    overlayOpacity : 0.75,
-    autoDimensions: false
-  });
-}
-
 $(document).ready(function() {
-
 
 });