+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Create variable to load modal : bootstrap, fancybox 1 or none
 * Modal compatible with Gravity Forms
 * Make a poll plugin compatible with TB
+* Fix gallery page
+* Add widgets from JetPack http://jetpack.me/support/extra-sidebar-widgets/
 
 ### 0.0.2: February 27th, 2013
 * New template for attachements